Move money commands in own package
